move rolescope.go to role/ pkg
